Package move only: move map-related and spatial classes into a new "geo" package.
