+++ collect_app/src/main/java/org/odk/collect/android/activities/GeoPolyActivity.java
@@ -29,9 +29,9 @@ import android.widget.Spinner;
 import android.widget.TextView;
 
 import org.odk.collect.android.R;
-import org.odk.collect.android.map.MapProvider;
-import org.odk.collect.android.map.MapFragment;
-import org.odk.collect.android.map.MapPoint;
+import org.odk.collect.android.geo.MapProvider;
+import org.odk.collect.android.geo.MapFragment;
+import org.odk.collect.android.geo.MapPoint;
 import org.odk.collect.android.preferences.MapsPreferences;
 import org.odk.collect.android.utilities.ToastUtils;
