+++ collect_app/src/main/java/org/odk/collect/android/geo/MapConfigurator.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.map;
+package org.odk.collect.android.geo;
 
 import android.content.Context;
 import android.content.SharedPreferences;
