+++ collect_app/src/main/java/org/odk/collect/android/geo/OsmMBTileProvider.java
@@ -16,7 +16,7 @@
  * @author Jon Nordling (jonnordling@gmail.com)
  */
 
-package org.odk.collect.android.spatial;
+package org.odk.collect.android.geo;
 
 import org.osmdroid.tileprovider.IRegisterReceiver;
 import org.osmdroid.tileprovider.MapTileProviderArray;
