+++ collect_app/src/test/java/org/odk/collect/android/location/activities/GeoPointMapActivityTest.java
@@ -8,7 +8,7 @@ import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 import org.odk.collect.android.R;
 import org.odk.collect.android.activities.GeoPointMapActivity;
-import org.odk.collect.android.map.MapPoint;
+import org.odk.collect.android.geo.MapPoint;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.android.controller.ActivityController;
