+++ collect_app/src/test/java/org/odk/collect/android/location/activities/GeoPolyActivityTest.java
@@ -21,7 +21,7 @@ import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 import org.odk.collect.android.activities.GeoPolyActivity;
-import org.odk.collect.android.map.MapPoint;
+import org.odk.collect.android.geo.MapPoint;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.android.controller.ActivityController;
